pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cyclist_census"
version = "0.0.1"
description = "This project is focused on performing inference with detection and classification models."
authors = [
  { name = "Juan Cruz Ferreyra" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "black",
    "flake8",
    "isort",
    "mkdocs",
    "pip",
    "pytest",
    "python-dotenv",
    "numpy (>=2,<2.3.0)",
    "supervision (==0.26.1)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "ultralytics (>=8.3.163,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1)",
    "albumentations (>=2.0.8,<3.0.0)",
    "mlflow (>=3.1.4,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "rfdetr (>=1.2.1,<2.0.0)",
    "trackers (>=2.0.1,<3.0.0)",
]
requires-python = "~=3.11.0"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["cyclist_census"]
force_sort_within_sections = true

[[tool.poetry.source]]
name = "pytorch_gpu"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"

[tool.poetry.dependencies]
torch = { source = "pytorch_gpu" }
torchvision = { source = "pytorch_gpu" }
